Moved menu state to component level for performance Changed menutoggle to opacity for performance added target tags on links so they open in new tab
Relates #109
