--- components/menuShared.js
@@ -1,294 +1,0 @@
-import React from "react"
-import styled from "styled-components"
-
-const Container = styled.div.attrs(({ showMenu }) => ({
-  className: `fixed top-0 left-0 vh-100 w-100 z-1 bg-white ${
-    showMenu ? "o-100" : "o-0 disabled-link"
-  }
-  `,
-}))``
-
-const Close = styled.img.attrs({
-  className: "h2 w2 absolute",
-})`
-  top: 6vh;
-  left: 5vw;
-`
-
-const MenuShared = ({
-  showMenu,
-  toggleMenu,
-  shareOnFacebook,
-  getTwitterHref,
-}) => (
-  <Container showMenu={showMenu}>
-    <Close src="../static/images/close_menu.svg" onClick={toggleMenu} />
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 375 667">
-      <g transform="translate(-4 -238)">
-        <polygon
-          fill="#FF62C6"
-          points="311.733 293.615 5.366 293.796 0 249.473 5.366 .695 318.049 .695"
-        />
-        <polygon
-          fill="#FAC744"
-          points="69.751 209.043 65.366 503.434 379.512 503.434 379.512 208.86"
-        />
-        <polygon
-          fill="#4B81EF"
-          points="4.39 557.439 296.585 557.439 296.585 262.866 4.39 262.866"
-        />
-        <polygon
-          fill="#FF62C6"
-          points="295.122 557.24 379.69 557.439 380 502.967 295.391 429.791"
-        />
-        <polygon
-          fill="#FF4B3A"
-          points="5.366 262.748 69.925 208.86 296.585 262.866"
-        />
-        <polygon
-          fill="#FAC744"
-          points="4.39 611.444 4.39 906.018 314.151 905.836 318.537 611.444"
-        />
-        <polygon
-          fill="#26302F"
-          points="87.317 852.013 379.512 852.013 379.512 557.439 87.317 557.439"
-        />
-        <polygon
-          fill="#4B81EF"
-          points="4.39 611.841 87.54 685.088 96.268 554.033 2.522 542.711"
-        />
-        <polygon
-          fill="#FF4B3A"
-          points="314.439 907 379.512 907 379.512 852.091 87.805 852.013"
-        />
-        <polygon
-          fill="#4B81EF"
-          points="87.317 699.817 231.707 699.817 231.707 556.948 87.317 556.948"
-        />
-        <polygon
-          fill="#FF62C6"
-          points="162.439 557.439 296.585 557.439 296.585 410.643 162.439 410.643"
-        />
-      </g>
-      <g fill="#FFF" transform="translate(166 175)" onClick={shareOnFacebook}>
-        <path d="M61.2405762,33 L64.452002,33 C64.5672363,33.0161965 64.6824707,33.0412274 64.7981934,33.0476078 C68.054541,33.2188977 71.311377,33.3833164 74.5672363,33.5624592 C75.1868652,33.5963245 75.8025879,33.6861413 76.4144043,33.7494547 L76.4144043,50.4273996 L75.932959,50.4273996 C72.957373,50.4273996 69.9812988,50.4224915 67.0057129,50.4293628 C65.6736816,50.4327984 64.3460449,50.4892406 63.0394043,50.7896115 C60.7825684,51.3088801 59.2669434,52.6114687 58.8465332,54.954067 C58.6438965,56.0834025 58.5516113,57.245131 58.5398926,58.3936078 C58.5008301,62.0672612 58.5218262,65.7404238 58.5213379,69.4135865 C58.5213379,69.5264709 58.5369629,69.6383738 58.5467285,69.7689272 L76.9065918,69.7689272 L76.9065918,70.2415695 C76.8836426,70.3151898 76.8484863,70.3878286 76.8396973,70.4629213 C76.6512207,72.1159426 76.4661621,73.7689639 76.2786621,75.4224761 C76.0311035,77.6094901 75.7825684,79.7960132 75.5345215,81.9825364 C75.285498,84.1818205 75.0374512,86.3820861 74.7864746,88.5808793 C74.7054199,89.2925423 74.6180176,90.0032237 74.5281738,90.7605313 L58.5203613,90.7605313 L58.5203613,91.3598006 C58.5203613,106.720923 58.5203613,122.081556 58.5193848,137.442678 C58.5193848,137.638999 58.5057129,137.83532 58.498877,138.031641 L37.6629395,138.031641 C37.6556152,137.83532 37.6429199,137.638508 37.6429199,137.442678 C37.6419434,122.07812 37.6419434,106.713561 37.6419434,91.3494937 L37.6419434,90.7683841 L22.0003418,90.7683841 L22.0003418,69.7605835 L37.6419434,69.7605835 L37.6419434,69.2457322 C37.6419434,65.4798079 37.6497559,61.7143745 37.639502,57.9484503 C37.6307129,54.9702634 37.9393066,52.0347763 38.7757324,49.1694739 C40.0057129,44.9534842 42.1780762,41.3348006 45.5560059,38.5009095 C49.498877,35.191922 54.1312012,33.6380427 59.176123,33.1600015 C59.863623,33.0947248 60.5525879,33.0525158 61.2405762,33" />
-        <text fontFamily="Apercu" fontSize="24">
-          <tspan x="5.96" y="21.224">
-            Share to
-          </tspan>
-        </text>
-      </g>
-      <g fill="#FFF" transform="translate(102 331)">
-        <path d="M40.0032766,117.98234 L36.6589895,117.98234 C36.5327695,117.966035 36.4070935,117.939403 36.2803294,117.935055 C32.3941667,117.813853 28.560777,117.29698 24.7785281,116.39911 C18.6829706,114.952299 12.9769542,112.567399 7.64742177,109.279737 C7.4287129,109.144404 7.21544455,108.999832 7,108.859064 C12.3953627,109.390068 17.6938842,109.016137 22.901005,107.559 C28.1026853,106.103493 32.8337606,103.715332 37.1546208,100.337992 C26.5042608,100.042868 19.8570346,92.2897687 18.1100839,86.1492485 C19.6209161,86.4324147 21.1350125,86.5541598 22.6572697,86.4916567 C24.1751746,86.4286101 25.67839,86.2438184 27.135905,85.7910787 C15.6602182,83.1692103 10.5765972,72.9540327 10.9329512,65.5058393 C13.7805189,67.0580898 16.8070796,67.9108492 20.1187236,68.0255288 C9.672927,60.8349568 9.28556205,48.0278035 13.8539659,40.7649454 C19.0774082,47.1054755 25.3242122,52.1758342 32.6515032,55.8972133 C39.9815145,59.6185925 47.7581924,61.6676066 55.9973145,62.1029541 C55.9652154,61.9360981 55.9434533,61.8214185 55.9200591,61.7067389 C55.3264985,58.8495328 55.3090888,55.9901527 55.9380128,53.1394686 C57.0152356,48.2609672 59.524947,44.2624005 63.4997902,41.2367078 C68.3445724,37.5473954 73.8199108,36.2918988 79.796321,37.3751043 C84.0486306,38.1457945 87.6709282,40.1632853 90.6860639,43.2596336 C90.8177244,43.3949664 90.9265348,43.4444254 91.12729,43.4036625 C95.309417,42.5557946 99.2586897,41.0823524 102.984901,39.00725 C103.259647,38.8550686 103.532761,38.7012567 103.883131,38.5050514 C102.240638,43.3759437 99.2543373,47.0674301 94.9155234,49.8050651 C99.0116902,49.2979749 102.848888,48.2446621 106.609375,46.6016463 C106.512534,46.7549147 106.46901,46.8277444 106.421677,46.8978565 C103.74875,50.817615 100.501304,54.1824549 96.6641062,56.9771579 C96.5101395,57.0885764 96.4062256,57.1842333 96.4181947,57.4092444 C96.5498553,59.8294724 96.4753202,62.2475265 96.2625959,64.6601454 C95.9862175,67.7956261 95.4525026,70.8898004 94.6690679,73.9394073 C93.2496364,79.4646798 91.1180411,84.6997205 88.2394625,89.6298549 C84.283117,96.4068195 79.2196259,102.202758 72.920593,106.90462 C67.3467812,111.066238 61.1897458,114.034319 54.4804978,115.864844 C52.0855813,116.518137 49.6618301,117.043707 47.1961868,117.343722 C45.4971127,117.550253 43.7898779,117.693739 42.0848192,117.850812 C41.3927852,117.914402 40.6969428,117.939947 40.0032766,117.98234" />
-        <text fontFamily="Apercu" fontSize="24">
-          <tspan x="5.96" y="21.224">
-            Share to
-          </tspan>
-        </text>
-      </g>
-      <g transform="translate(78 45)">
-        <rect
-          width="92.895"
-          height="36.393"
-          x="10.867"
-          y=".645"
-          fill="#FF62C6"
-        />
-        <text
-          fill="#FFF"
-          fontFamily="Apercu"
-          fontSize="32"
-          textDecoration="underline"
-        >
-          <tspan x="10.84" y="28.632">
-            Home
-          </tspan>
-        </text>
-      </g>
-      <g transform="translate(89 83)">
-        <rect width="108.543" height="32.393" y="3" fill="#FF62C6" />
-        <text
-          fill="#FFF"
-          fontFamily="Apercu"
-          fontSize="32"
-          textDecoration="underline"
-        >
-          <tspan x="1" y="30">
-            Privacy
-          </tspan>
-        </text>
-      </g>
-      <g transform="translate(68 121)">
-        <rect width="63.246" height="32.393" x="21" y="2" fill="#FF62C6" />
-        <text
-          fill="#FFF"
-          fontFamily="Apercu"
-          fontSize="32"
-          textDecoration="underline"
-        >
-          <tspan x="19.693" y="27.632">
-            FAQ
-          </tspan>
-        </text>
-      </g>
-    </svg>
-  </Container>
-)
-
-export default MenuShared
-
-// const CloseButton = ({ closeOnClick }) => (
-//   <g
-//     id="close_btn"
-//     onTouchStart={closeOnClick}
-//     transform="translate(38.000000, 45.000000)"
-//     stroke="#FFFFFF"
-//     strokeLinecap="round"
-//     strokeWidth="7"
-//   >
-//     <path d="M0,0 L24.5192308,25.5" id="Path-2" />
-//     <path d="M0.138940912,25.7149429 L24.7229721,0.147550558" id="Path-3" />
-//   </g>
-// )
-
-// const FacebookButton = ({ fbOnClick }) => (
-//   <g
-//     id="fb_link"
-//     transform="translate(166.000000, 175.000000)"
-//     onClick={fbOnClick}
-//   >
-//     <g id="fb_icon" transform="translate(22.000000, 33.000000)">
-//       <mask id="mask-2" fill="white">
-//         <use xlinkHref="#path-1" />
-//       </mask>
-//       <g id="Clip-2" />
-//       <path
-//         d="M39.2405762,0 L42.452002,0 C42.5672363,0.016196468 42.6824707,0.0412273731 42.7981934,0.0476077999 C46.054541,0.218897719 49.311377,0.383316409 52.5672363,0.562459161 C53.1868652,0.596324503 53.8025879,0.68614128 54.4144043,0.749454746 L54.4144043,17.4273996 L53.932959,17.4273996 C50.957373,17.4273996 47.9812988,17.4224915 45.0057129,17.4293628 C43.6736816,17.4327984 42.3460449,17.4892406 41.0394043,17.7896115 C38.7825684,18.3088801 37.2669434,19.6114687 36.8465332,21.954067 C36.6438965,23.0834025 36.5516113,24.245131 36.5398926,25.3936078 C36.5008301,29.0672612 36.5218262,32.7404238 36.5213379,36.4135865 C36.5213379,36.5264709 36.5369629,36.6383738 36.5467285,36.7689272 L54.9065918,36.7689272 L54.9065918,37.2415695 C54.8836426,37.3151898 54.8484863,37.3878286 54.8396973,37.4629213 C54.6512207,39.1159426 54.4661621,40.7689639 54.2786621,42.4224761 C54.0311035,44.6094901 53.7825684,46.7960132 53.5345215,48.9825364 C53.285498,51.1818205 53.0374512,53.3820861 52.7864746,55.5808793 C52.7054199,56.2925423 52.6180176,57.0032237 52.5281738,57.7605313 L36.5203613,57.7605313 L36.5203613,58.3598006 C36.5203613,73.7209235 36.5203613,89.0815556 36.5193848,104.442678 C36.5193848,104.638999 36.5057129,104.83532 36.498877,105.031641 L15.6629395,105.031641 C15.6556152,104.83532 15.6429199,104.638508 15.6429199,104.442678 C15.6419434,89.0781199 15.6419434,73.7135614 15.6419434,58.3494937 L15.6419434,57.7683841 L0.000341796875,57.7683841 L0.000341796875,36.7605835 L15.6419434,36.7605835 L15.6419434,36.2457322 C15.6419434,32.4798079 15.6497559,28.7143745 15.639502,24.9484503 C15.6307129,21.9702634 15.9393066,19.0347763 16.7757324,16.1694739 C18.0057129,11.9534842 20.1780762,8.33480059 23.5560059,5.50090949 C27.498877,2.191922 32.1312012,0.638042678 37.176123,0.160001472 C37.863623,0.0947247976 38.5525879,0.0525158205 39.2405762,0"
-//         id="Fill-1"
-//         fill="#FFFFFF"
-//         mask="url(#mask-2)"
-//       />
-//     </g>
-//     <text
-//       id="Share-to"
-//       fontFamily="Apercu"
-//       fontSize="24"
-//       fontWeight="normal"
-//       line-spacing="23"
-//       fill="#FFFFFF"
-//     >
-//       <tspan x="5.96020749" y="21.224">
-//         Share to
-//       </tspan>
-//     </text>
-//   </g>
-// )
-// const TwitterButton = ({ twitterHref }) => (
-//   <a href={twitterHref} alt="Share on Twitter">
-//     <g id="twitter_link" transform="translate(102.000000, 331.000000)">
-//       <g id="twitter_icon" transform="translate(7.000000, 37.000000)">
-//         <mask id="mask-4" fill="white">
-//           <use xlinkHref="#path-3" />
-//         </mask>
-//         <g id="Clip-2" />
-//         <path
-//           d="M33.0032766,80.98234 L29.6589895,80.98234 C29.5327695,80.9660348 29.4070935,80.9394031 29.2803294,80.935055 C25.3941667,80.8138534 21.560777,80.2969801 17.7785281,79.3991097 C11.6829706,77.9522992 5.97695422,75.567399 0.647421766,72.2797369 C0.4287129,72.1444042 0.215444554,71.9998318 0,71.859064 C5.39536273,72.3900684 10.6938842,72.016137 15.901005,70.5589999 C21.1026853,69.1034933 25.8337606,66.7153321 30.1546208,63.3379917 C19.5042608,63.0428684 12.8570346,55.2897687 11.1100839,49.1492485 C12.6209161,49.4324147 14.1350125,49.5541598 15.6572697,49.4916567 C17.1751746,49.4286101 18.67839,49.2438184 20.135905,48.7910787 C8.66021821,46.1692103 3.57659722,35.9540327 3.93295121,28.5058393 C6.78051888,30.0580898 9.80707962,30.9108492 13.1187236,31.0255288 C2.672927,23.8349568 2.28556205,11.0278035 6.85396589,3.76494541 C12.0774082,10.1054755 18.3242122,15.1758342 25.6515032,18.8972133 C32.9815145,22.6185925 40.7581924,24.6676066 48.9973145,25.1029541 C48.9652154,24.9360981 48.9434533,24.8214185 48.9200591,24.7067389 C48.3264985,21.8495328 48.3090888,18.9901527 48.9380128,16.1394686 C50.0152356,11.2609672 52.524947,7.26240048 56.4997902,4.2367078 C61.3445724,0.547395449 66.8199108,-0.708101242 72.796321,0.375104345 C77.0486306,1.14579452 80.6709282,3.16328531 83.6860639,6.25963364 C83.8177244,6.3949664 83.9265348,6.44442536 84.12729,6.40366248 C88.309417,5.55579459 92.2586897,4.08235237 95.9849012,2.00725004 C96.2596474,1.85506863 96.5327615,1.70125669 96.8831309,1.50505137 C95.2406382,6.37594372 92.2543373,10.0674301 87.9155234,12.8050651 C92.0116902,12.2979749 95.8488882,11.2446621 99.609375,9.60164626 C99.5125338,9.75491469 99.4690096,9.82774437 99.4216771,9.89785652 C96.7487501,13.817615 93.5013043,17.1824549 89.6641062,19.9771579 C89.5101395,20.0885764 89.4062256,20.1842333 89.4181947,20.4092444 C89.5498553,22.8294724 89.4753202,25.2475265 89.2625959,27.6601454 C88.9862175,30.7956261 88.4525026,33.8898004 87.6690679,36.9394073 C86.2496364,42.4646798 84.1180411,47.6997205 81.2394625,52.6298549 C77.283117,59.4068195 72.2196259,65.2027575 65.920593,69.9046198 C60.3467812,74.0662381 54.1897458,77.0343192 47.4804978,78.8648442 C45.0855813,79.5181373 42.6618301,80.0437067 40.1961868,80.3437215 C38.4971127,80.5502534 36.7898779,80.6937388 35.0848192,80.8508117 C34.3927852,80.9144018 33.6969428,80.9399466 33.0032766,80.98234"
-//           id="Fill-1"
-//           fill="#FFFFFF"
-//           mask="url(#mask-4)"
-//         />
-//       </g>
-//       <text
-//         id="Share-to"
-//         fontFamily="Apercu"
-//         fontSize="24"
-//         fontWeight="normal"
-//         line-spacing="23"
-//         fill="#FFFFFF"
-//       >
-//         <tspan x="5.96020749" y="21.224">
-//           Share to
-//         </tspan>
-//       </text>
-//     </g>
-//   </a>
-// )
-// const NavLinks = () => (
-//   <g id="navLinks">
-//     <a href="https://createrefresh.eu/" alt="Home">
-//       <g id="home_link" transform="translate(78.000000, 45.000000)">
-//         <rect
-//           id="home_bg"
-//           fill="#FF62C6"
-//           x="10.8671875"
-//           y="0.644591611"
-//           width="92.8945312"
-//           height="36.392936"
-//         />
-//         <text
-//           id="Home"
-//           fontFamily="Apercu"
-//           fontSize="32"
-//           fontWeight="normal"
-//           line-spacing="31"
-//           fill="#FFFFFF"
-//           textDecoration="underline"
-//         >
-//           <tspan x="10.8402075" y="28.632">
-//             Home
-//           </tspan>
-//         </text>
-//       </g>
-//     </a>
-//     <a href="https://createrefresh.eu/privacy/" alt="Privacy">
-//       <g id="privacy_link" transform="translate(89.000000, 83.000000)">
-//         <rect
-//           id="privacy_bg"
-//           fill="#FF62C6"
-//           x="0"
-//           y="3"
-//           width="108.542969"
-//           height="32.392936"
-//         />
-//         <text
-//           id="Privacy"
-//           fontFamily="Apercu"
-//           fontSize="32"
-//           fontWeight="normal"
-//           line-spacing="31"
-//           fill="#FFFFFF"
-//           textDecoration="underline"
-//         >
-//           <tspan x="1" y="30">
-//             Privacy
-//           </tspan>
-//         </text>
-//       </g>
-//     </a>
-//     <a href="https://createrefresh.eu/faq/" alt="FAQs">
-//       <g id="faq_link" transform="translate(68.000000, 121.000000)">
-//         <rect
-//           id="faq_bg"
-//           fill="#FF62C6"
-//           x="21"
-//           y="2"
-//           width="63.2460938"
-//           height="32.392936"
-//         />
-//         <text
-//           id="FAQ"
-//           fontFamily="Apercu"
-//           fontSize="32"
-//           fontWeight="normal"
-//           line-spacing="31"
-//           fill="#FFFFFF"
-//           textDecoration="underline"
-//         >
-//           <tspan x="19.6934434" y="27.632">
-//             FAQ
-//           </tspan>
-//         </text>
-//       </g>
-//     </a>
-//   </g>
-// )
