+++ pages/index.js
@@ -30,7 +30,6 @@ class IndexPage extends Component {
     cartoon: null,
     cartoonId: null,
     view: "",
-    showMenu: false,
     showShareModal: false,
     formCompleted: false,
   }
@@ -163,15 +162,6 @@ class IndexPage extends Component {
     this.setState({ view: "loading" })
   }
 
-  toggleMenu = () => {
-    this.setState({ showMenu: !this.state.showMenu })
-    if (this.state.showMenu) {
-      document.body.style.position = "relative"
-    } else {
-      document.body.style.position = "fixed"
-    }
-  }
-
   toggleShareModal = () => {
     this.setState(prevProps => ({ showShareModal: !prevProps.showShareModal }))
   }
@@ -209,14 +199,12 @@ class IndexPage extends Component {
           error={error}
           cartoon={cartoon}
           view={view}
-          showMenu={showMenu}
           showShareModal={showShareModal}
           submitForm={this.submitForm}
           seeMeme={this.seeMeme}
           onImageSelect={this.onImageSelect}
           handleStartOver={this.handleStartOver}
           toggleShareModal={this.toggleShareModal}
-          toggleMenu={this.toggleMenu}
         />
         <Info />
         <Video />
